moved data types from utils to data_types
